+++ packages/keyring-api/src/api/account.ts
@@ -1,18 +1,9 @@
 import { AccountIdStruct, object } from '@metamask/keyring-utils';
 import type { Infer } from '@metamask/superstruct';
-import {
-  nonempty,
-  array,
-  enums,
-  record,
-  string,
-  union,
-} from '@metamask/superstruct';
-import {
-  CaipChainIdStruct,
-  CaipNamespaceStruct,
-  JsonStruct,
-} from '@metamask/utils';
+import { nonempty, array, enums, record, string } from '@metamask/superstruct';
+import { JsonStruct } from '@metamask/utils';
+
+import { CaipChainIdStruct } from './caip';
 
 /**
  * Supported Ethereum account types.
@@ -74,9 +65,9 @@ export const KeyringAccountStruct = object({
   address: string(),
 
   /**
-   * Account supported scopes (CAIP-2 chain IDs or CAIP-2 namespaces).
+   * Account supported scopes (CAIP-2 chain IDs).
    */
-  scopes: nonempty(array(union([CaipNamespaceStruct, CaipChainIdStruct]))),
+  scopes: nonempty(array(CaipChainIdStruct)),
 
   /**
    * Account options.