+++ packages/keyring-api/src/sol/types.ts
@@ -1,9 +1,13 @@
-import { object, definePattern } from '@metamask/keyring-utils';
+import { object } from '@metamask/keyring-utils';
 import type { Infer } from '@metamask/superstruct';
 import { array, enums, literal, nonempty } from '@metamask/superstruct';
-import { CaipChainIdStruct } from '@metamask/utils';
+import { definePattern } from '@metamask/utils';
 
-import { KeyringAccountStruct, SolAccountType } from '../api';
+import {
+  CaipChainIdStruct,
+  KeyringAccountStruct,
+  SolAccountType,
+} from '../api';
 
 /**
  * Solana addresses are represented in the format of a 256-bit ed25519 public key and