+++ packages/keyring-snap-bridge/src/logger.ts
@@ -1,4 +1,4 @@
-/* istanbul ignore file */
+// istanbul ignore file
 
 import { createProjectLogger, createModuleLogger } from '@metamask/utils';