+++ packages/keyring-snap-client/src/KeyringClient.test.ts
@@ -1,11 +1,13 @@
+import { BtcMethod, KeyringRpcMethod } from '@metamask/keyring-api';
 import type {
-  CaipAssetType,
-  CaipAssetTypeOrId,
   KeyringAccount,
   KeyringRequest,
   KeyringResponse,
+  CaipChainId,
+  CaipAssetType,
+  CaipAssetTypeOrId,
 } from '@metamask/keyring-api';
-import { KeyringRpcMethod } from '@metamask/keyring-api';
+import type { JsonRpcRequest } from '@metamask/keyring-utils';
 
 import { KeyringClient } from '.'; // Import from `index.ts` to test the public API
 
@@ -28,7 +30,7 @@ describe('KeyringClient', () => {
           address: '0xE9A74AACd7df8112911ca93260fC5a046f8a64Ae',
           options: {},
           methods: [],
-          scopes: ['eip155'],
+          scopes: ['eip155:0'],
           type: 'eip155:eoa',
         },
       ];
@@ -52,7 +54,7 @@ describe('KeyringClient', () => {
         address: '0xE9A74AACd7df8112911ca93260fC5a046f8a64Ae',
         options: {},
         methods: [],
-        scopes: ['eip155'],
+        scopes: ['eip155:0'],
         type: 'eip155:eoa',
       };
 
@@ -75,7 +77,7 @@ describe('KeyringClient', () => {
         address: '0xE9A74AACd7df8112911ca93260fC5a046f8a64Ae',
         options: {},
         methods: [],
-        scopes: ['eip155'],
+        scopes: ['eip155:0'],
         type: 'eip155:eoa',
       };
 
@@ -411,6 +413,70 @@ describe('KeyringClient', () => {
     });
   });
 
+  describe('resolveAccountAddress', () => {
+    const scope: CaipChainId = 'bip122:000000000019d6689c085ae165831e93';
+    const request: JsonRpcRequest = {
+      id: '71621d8d-62a4-4bf4-97cc-fb8f243679b0',
+      jsonrpc: '2.0',
+      method: BtcMethod.SendBitcoin,
+      params: {
+        recipients: {
+          address: '0.1',
+        },
+        replaceable: true,
+      },
+    };
+
+    it('should send a request to resolve an account address from a signing request and return the response', async () => {
+      const expectedResponse = {
+        address: 'tb1qspc3kwj3zfnltjpucn7ugahr8lhrgg86wzpvs3',
+      };
+
+      mockSender.send.mockResolvedValue(expectedResponse);
+      const account = await keyring.resolveAccountAddress(scope, request);
+      expect(mockSender.send).toHaveBeenCalledWith({
+        jsonrpc: '2.0',
+        id: expect.any(String),
+        method: 'keyring_resolveAccountAddress',
+        params: {
+          scope,
+          request,
+        },
+      });
+      expect(account).toStrictEqual(expectedResponse);
+    });
+
+    it('should send a request to resolve an account address from a signing request and return null if the address cannot be resolved', async () => {
+      const expectedResponse = null;
+
+      mockSender.send.mockResolvedValue(expectedResponse);
+      const account = await keyring.resolveAccountAddress(scope, request);
+      expect(mockSender.send).toHaveBeenCalledWith({
+        jsonrpc: '2.0',
+        id: expect.any(String),
+        method: 'keyring_resolveAccountAddress',
+        params: {
+          scope,
+          request,
+        },
+      });
+      expect(account).toStrictEqual(expectedResponse);
+    });
+
+    it('should throw an exception if the response is malformed', async () => {
+      const expectedResponse = {
+        not: 'good',
+      };
+
+      mockSender.send.mockResolvedValue(expectedResponse);
+      await expect(
+        keyring.resolveAccountAddress(scope, request),
+      ).rejects.toThrow(
+        'At path: address -- Expected a string, but received: undefined',
+      );
+    });
+  });
+
   describe('filterAccountChains', () => {
     it('should send a request to filter the chains supported by an account and return the response', async () => {
       const id = '49116980-0712-4fa5-b045-e4294f1d440e';
@@ -438,7 +504,7 @@ describe('KeyringClient', () => {
         address: '0xE9A74AACd7df8112911ca93260fC5a046f8a64Ae',
         options: {},
         methods: [],
-        scopes: ['eip155'],
+        scopes: ['eip155:0'],
         type: 'eip155:eoa',
       };