+++ packages/keyring-snap-sdk/src/rpc-handler.test.ts
@@ -1,4 +1,8 @@
-import { KeyringRpcMethod, isKeyringRpcMethod } from '@metamask/keyring-api';
+import {
+  BtcMethod,
+  KeyringRpcMethod,
+  isKeyringRpcMethod,
+} from '@metamask/keyring-api';
 import type { Keyring, GetAccountBalancesRequest } from '@metamask/keyring-api';
 import type { JsonRpcRequest } from '@metamask/keyring-utils';
 
@@ -12,6 +16,7 @@ describe('handleKeyringRequest', () => {
     listAccountTransactions: jest.fn(),
     listAccountAssets: jest.fn(),
     getAccountBalances: jest.fn(),
+    resolveAccountAddress: jest.fn(),
     filterAccountChains: jest.fn(),
     updateAccount: jest.fn(),
     deleteAccount: jest.fn(),
@@ -192,6 +197,72 @@ describe('handleKeyringRequest', () => {
     );
   });
 
+  it('calls `keyring_resolveAccountAddress`', async () => {
+    const scope = 'bip122:000000000019d6689c085ae165831e93';
+    const signingRequest = {
+      id: '71621d8d-62a4-4bf4-97cc-fb8f243679b0',
+      jsonrpc: '2.0',
+      method: BtcMethod.SendBitcoin,
+      params: {
+        recipients: {
+          address: '0.1',
+        },
+        replaceable: true,
+      },
+    };
+    const request: JsonRpcRequest = {
+      jsonrpc: '2.0',
+      id: '7c507ff0-365f-4de0-8cd5-eb83c30ebda4',
+      method: 'keyring_resolveAccountAddress',
+      params: {
+        scope,
+        request: signingRequest,
+      },
+    };
+
+    keyring.resolveAccountAddress.mockResolvedValue(
+      'ResolveAccountAddress result',
+    );
+    const result = await handleKeyringRequest(keyring, request);
+
+    expect(keyring.resolveAccountAddress).toHaveBeenCalledWith(
+      scope,
+      signingRequest,
+    );
+    expect(result).toBe('ResolveAccountAddress result');
+  });
+
+  it('throws an error if `keyring_resolveAccountAddress` is not implemented', async () => {
+    const scope = 'bip122:000000000019d6689c085ae165831e93';
+    const signingRequest = {
+      id: '71621d8d-62a4-4bf4-97cc-fb8f243679b0',
+      jsonrpc: '2.0',
+      method: BtcMethod.SendBitcoin,
+      params: {
+        recipients: {
+          address: '0.1',
+        },
+        replaceable: true,
+      },
+    };
+    const request: JsonRpcRequest = {
+      jsonrpc: '2.0',
+      id: '7c507ff0-365f-4de0-8cd5-eb83c30ebda4',
+      method: 'keyring_resolveAccountAddress',
+      params: {
+        scope,
+        request: signingRequest,
+      },
+    };
+
+    const partialKeyring: Keyring = { ...keyring };
+    delete partialKeyring.resolveAccountAddress;
+
+    await expect(handleKeyringRequest(partialKeyring, request)).rejects.toThrow(
+      'Method not supported: keyring_resolveAccountAddress',
+    );
+  });
+
   it('calls `keyring_filterAccountChains`', async () => {
     const request: JsonRpcRequest = {
       jsonrpc: '2.0',
@@ -226,7 +297,7 @@ describe('handleKeyringRequest', () => {
           address: '0x0',
           options: {},
           methods: [],
-          scopes: ['eip155'],
+          scopes: ['eip155:0'],
           type: 'eip155:eoa',
         },
       },
@@ -240,7 +311,7 @@ describe('handleKeyringRequest', () => {
       address: '0x0',
       options: {},
       methods: [],
-      scopes: ['eip155'],
+      scopes: ['eip155:0'],
       type: 'eip155:eoa',
     });
     expect(result).toBe('UpdateAccount result');