+++ packages/keyring-snap-sdk/src/rpc-handler.ts
@@ -16,6 +16,7 @@ import {
   ListRequestsRequestStruct,
   GetAccountBalancesRequestStruct,
   ListAccountAssetsRequestStruct,
+  ResolveAccountAddressRequestStruct,
 } from '@metamask/keyring-api';
 import type { JsonRpcRequest } from '@metamask/keyring-utils';
 import { JsonRpcRequestStruct } from '@metamask/keyring-utils';
@@ -90,6 +91,17 @@ async function dispatchRequest(
       return keyring.getAccountBalances(
         request.params.id,
         request.params.assets,
+      );
+    }
+
+    case `${KeyringRpcMethod.ResolveAccountAddress}`: {
+      if (keyring.resolveAccountAddress === undefined) {
+        throw new MethodNotSupportedError(request.method);
+      }
+      assert(request, ResolveAccountAddressRequestStruct);
+      return keyring.resolveAccountAddress(
+        request.params.scope,
+        request.params.request,
       );
     }